requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-binance==1.0.16
pandas==1.3.3
numpy==1.21.2
TA-Lib==0.4.24
requests==2.26.0
python-telegram-bot==13.7
schedule==1.1.0
configparser==5.0.2
